refactor: move community list component at page level + SEO improvements on pagination
