+++ src/app/(index)/components/add-community-button.tsx
